fast_hadamard_a5: merge the two hadamard test files, correct a stale README
test_hadamard256_a5.py (54) and test_block_sizes_a5.py (91) were the same test
twice: identical Sylvester hadamard_matrix(), identical TOLERANCE, identical
importorskip preamble, and the same compare-to-reference body differing only in
which axis it swept. They are now one test_hadamard_a5.py (105) with one shared
assert_matches_reference(kernel, n, batch, seed, label).

Coverage is unchanged, and that was verified rather than assumed: collecting the
two old files from git next to the merged one gives 18 cases before and 18 after,
with an identical set of test names and parameters. Every seed and assertion
message format is preserved -- the batch sweep still seeds on batch, the N sweep
on n, the padding test on 1000+n.

The README was not merely verbose, it was stale: it still listed
jit_util_hadamard256_a5.py, jit_util_copy256_a5.py, test_hadamard256_a5.py and
test_block_sizes_a5.py, all removed in earlier commits, plus the old Cfg name.
File list corrected and cut to one entry per file.

benchmark.py now imports N and rows_for from jit_util_a5 instead of re-deriving
the 32 KB-tile rule inline, and nbuf_for() becomes cfg(rows) returning the
(nbuf, prefetch) pair that always travelled together. Both verified equivalent to
the old expressions at every swept n and every swept rows. Line count is flat at
206 -- black re-expanded what the squeeze saved -- so this is a dedup, not a cut.

Device-verified: 24/24 from the kernel directory and 24/24 from the repo root,
--nsweep across N=32..2048 and the grid both exit 0.
